Improve some behavior of find
- Change seedSearchStringFromSelection to "selection". This way it isn't
automatically seeded with the word under the cursor. I'm not sure which
way I'll like this better, but one thing I don't like about the default
is that it can contaminate search history with unwanted things.

- Turn cursorMoveOnType to false. The main reason for this change is
that the default of true makes it not work to return to the location
before the search; changing it to false allows this return (with
control-hyphen).