Change editor.formatOnPaste to false for Fortran
For C++, this is a helpful setting to have true: it indents properly
after pasting, whereas if this is false, it does *not* indent properly
after pasting. But for Fortran, having this set to true actually *stops*
it from indenting properly after pasting: having it false allows it to
indent properly after pasting. So I'm going to have this true in
general, but false for Fortran.