Disable automatic Copilot suggestions; trigger with alt+cmd+space
I'm setting `github.copilot.editor.enableAutoCompletions` to false: when
this is true, it's very in-your-face, which - at least initially - I'm
finding distracting. It seems like I get the same results from disabling
this and then doing a manual trigger.

Then using a more ergonomic keybinding to trigger Copilot suggestions,
with similar keybindings to accept a suggestion.