Change "when" conditions for accepting inline suggestions
I had started from the conditions for tab completion. But I found that
this was preventing the use of alt+cmd+space to trigger inline
suggestions when typing a git commit message (though it didn't prevent
it in code). It seems like we may only want this to work for
inlineSuggestionVisible, not inlineEditIsVisible, anyway, so I'm
changing this condition; this allows alt+cmd+space to be used for
triggering suggestions in git commit messages now.