Go back to standard keybindings for find next/previous
I was finding that it was taking too much mental energy to do find
next/previous in various applications because I wasn't 100% consistently
using these VSCode bindings of cmd+h for next and cmd+g for previous.
This may have been partly due to my use of my personal laptop where I
didn't have this set up system-wide, but I think it was at least partly
because of applications where I couldn't get this change supported. So I
decided that it's worth the slightly less ergonomic-ness of the standard
keybindings for these actions in order to have consistency and thus
better muscle memory that doesn't require thought.

Then, rather than using shift to find with selection, I'm using ctrl for
that purpose (so it can be combined with shift+cmd+g to create
shift+ctrl+cmd+g).

And then using cmd+h for setting selection anchor (which is nice and
ergonomic).

And then changing or getting rid of some other keybindings accordingly.