Fix "when" condition for triggering inline suggestions
The trigger for inline suggestions had stopped working. I noticed that
the out-of-the-box setting no longer had the
config.github.copilot.inlineSuggest.enable condition, so I removed it
here. (It may be that this has changed to something like
editor.inlineSuggest.enable, or it may be that this is gone entirely.)